refactor: client impl package
